create modules, use relative imports, update tests
