+++ api/src/engines/UserEngine/IUserEngine.ts
@@ -1,4 +1,4 @@
-import { UserDto } from 'src/infrastructure/dto/UserDto';
+import { UserDto } from '../../infrastructure/dto/UserDto';
 
 export const IUserEngineProvider = 'IUserEngine';
 export class IUserEngine {
