+++ api/src/main.ts
@@ -2,10 +2,10 @@ import * as session from 'express-session';
 import * as Redis from 'redis';
 import * as connectRedis from 'connect-redis';
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { ApiModule } from './api.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create(ApiModule);
   app.setGlobalPrefix('api');
 
   // TODO: load config from environment
