+++ api/src/modules/AppModule.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { AppController } from '../controllers/app/app.controller';
+import { AppManager } from '../managers/AppManager/AppManager';
+import { IAppManagerProvider } from '../managers/AppManager/IAppManager';
+
+@Module({
+  controllers: [AppController],
+  providers: [{ provide: IAppManagerProvider, useClass: AppManager }],
+})
+export class AppModule {}
