+++ api/src/modules/PostModule.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { IPostHandlerProvider } from '../handlers/PostHandler/IPostHandler';
+import { PostHandler } from '../handlers/PostHandler/PostHandler';
+
+@Module({
+  providers: [{ provide: IPostHandlerProvider, useClass: PostHandler }],
+})
+export class PostModule {}
