+++ api/src/modules/SessionModule.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { SessionController } from '../controllers/session/session.controller';
+
+@Module({
+  controllers: [SessionController],
+})
+export class SessionModule {}
